requirements.txt
 bottle
 pyyaml